Fix Swiper 관련된 StyledComponents -> StyledSwiper 파일로 분리
